+++ src/main/java/stairwaytoheaven/mobs/DoctorHumanMob.java
@@ -124,7 +124,7 @@ public class DoctorHumanMob extends HumanShop {
     public List<InventoryItem> getRecruitItems(ServerClient client) {
         GameRandom random = new GameRandom((long) this.getSettlerSeed() * 367L);
         return Collections.singletonList(
-                new InventoryItem("coin", random.getIntBetween(900, 1400)));
+                new InventoryItem("coin", random.getIntBetween(400, 700)));
     }
 
     @Override